+++ indexer/integration_test_util/types/db/transaction.go
@@ -6,7 +6,7 @@ type TransactionRecord struct {
 	ChainId     string
 	Height      int64
 	Hash        string
-	PartitionId int
+	PartitionId string
 
 	Epoch        int64
 	MessageTypes []string