+++ CHANGELOG.md
@@ -94,6 +94,10 @@ Templates for Unreleased:
 - (txs) [#56](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/56) Simple translate EVM event logs
 - (txs) [#59](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/59) Mapping contracts name and translate events
 - (indexer) [#64](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/64) Whitelist and Ignore-list for indexer
+- (sql) [#70](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/70) Remove record from `partition_table_info` when drop corresponding partitioned table.
+- (sql) [#71](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/71) Re-partition table `transaction`, use combination of epoch week + chain-id as partition-id
+- (sql) [#72](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/72) Add new column to allow configures number of weeks keep recent txs
+- (sql) [#74](https://github.com/bcdevtools/dymension-rollapp-block-explorer/pull/74) Add helper method to list prune-able partitioned tables of `transaction` table
 
 ### Bug Fixes